fastapi-backend/mqtt_fastapi_project/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
paho-mqtt==1.6.1
pydantic==1.10.13
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4